requirements.txt
streamlit
tensorflow
Pillow
numpy
pandas
plotly
